In BGA log parser, refuse to parse non-Ark Nova logs (#304)
Does what it says on the tin.

It turns out this is harder to do than you'd think. BGA Ark Nova logs
don't contain any variant of the game name in them ("ark nova",
"ark_nova", "arknova") or the game ID (1741).

So we have to infer this indirectly instead. We do that by looking to
see if in this game, there was a game-end scoring draw action in the
first few turns. That should be enough to filter out the vast majority
of games, while not being too susceptible to changes on BGA's side.

Fixes #301.